Separate App and World
