+++ projects/TestApplication/TestApplication.vcxproj
@@ -88,10 +88,13 @@
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\mesh_splatting.cpp" />
     <ClCompile Include="src\unit_tests.cpp" />
+    <ClCompile Include="src\world1.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\app.h" />
     <ClInclude Include="src\mesh_splatting.h" />
+    <ClInclude Include="src\world.h" />
+    <ClInclude Include="src\world1.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
