+++ projects/TestApplication/src/app.h
@@ -1,16 +1,9 @@
 #pragma once
 
-#include "mesh_splatting.h"
-
 #include "core/win/windows_application.h"
 #include "render/renderer.h"
 #include "world/scene.h"
 #include "world/camera.h"
-#include "world/gpu_resource_asset.h"
-
-#include <vector>
-
-class StaticMesh;
 
 struct AppState
 {
@@ -33,19 +26,11 @@ protected:
 	virtual void onWindowResize(uint32 newWidth, uint32 newHeight) override;
 
 private:
-	void createResources();
-	void destroyResources();
-
-private:
 	Scene scene;
 	Camera camera;
 	AppState appState;
 
-	MeshSplatting meshSplatting;
-	
-	StaticMesh* pbrtMesh = nullptr;
-	StaticMesh* ground = nullptr;
-	StaticMesh* wallA = nullptr;
+	class World* world = nullptr;
 
 	bool bViewportNeedsResize = false;
 	uint32 newViewportWidth = 0;
